Improve closing of ship connection
When announcing closing of a connection is ignored by the remote device then the connection continued to be active.

This change tells the remote device that is has a max of 500ms to close the connection and will close the connection after this time automatically as well.

Fixes https://github.com/enbility/ship-go/issues/32